# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manager API for ShopManagement",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/clients": {
            "get": {
                "tags": [
                    "AllClientsWithOptionalFilters"
                ],
                "summary": "Get all Clients with optional filters",
                "description": "Récupération de tous les clients avec possibilité de filtrer les clients qui ont des comptes, qui n'ont pas de compte. Ceux dont leurs compte est actif ou inactif",
                "operationId": "GetAllClientsWithOptionalFilters",
                "parameters": [
                    {
                        "name": "comptes",
                        "in": "query",
                        "description": "Filterer les clients par comptes (oui ou non)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "oui",
                                "non"
                            ]
                        }
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filterer les clients par etat (oui ou non)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "oui",
                                "non"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération reussie",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "StoreClient"
                ],
                "summary": "Store Client",
                "description": "Création d'un client avec la possibilité de lui créer un compte",
                "operationId": "StoreClient",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "surname",
                                    "telephone"
                                ],
                                "properties": {
                                    "surname": {
                                        "type": "string",
                                        "example": "bobou"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "761434522"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "namek"
                                    },
                                    "user": {
                                        "properties": {
                                            "nom": {
                                                "type": "string",
                                                "example": "Mahmoud"
                                            },
                                            "prenom": {
                                                "type": "string",
                                                "example": "Mahmoud"
                                            },
                                            "login": {
                                                "type": "string",
                                                "example": "Mahmoudine"
                                            },
                                            "photo": {
                                                "type": "string",
                                                "example": "https://images.pexels.com/photos/633432/pexels-photo-633432.jpeg?auto=compress&cs=tinysrgb&w=600"
                                            },
                                            "password": {
                                                "type": "string",
                                                "format": "password",
                                                "example": "Passer@123"
                                            },
                                            "password_confirmation": {
                                                "type": "string",
                                                "format": "password",
                                                "example": "Passer@123"
                                            },
                                            "role": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "3"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "surname",
                                    "telephone"
                                ],
                                "properties": {
                                    "surname": {
                                        "type": "string",
                                        "example": "bobou"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "761434522"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "namek"
                                    },
                                    "user": {
                                        "properties": {
                                            "nom": {
                                                "type": "string",
                                                "example": "Mahmoud"
                                            },
                                            "prenom": {
                                                "type": "string",
                                                "example": "Mahmoud"
                                            },
                                            "login": {
                                                "type": "string",
                                                "example": "Mahmoudine"
                                            },
                                            "photo": {
                                                "type": "string",
                                                "example": "https://images.pexels.com/photos/633432/pexels-photo-633432.jpeg?auto=compress&cs=tinysrgb&w=600"
                                            },
                                            "password": {
                                                "type": "string",
                                                "format": "password",
                                                "example": "Passer@123"
                                            },
                                            "password_confirmation": {
                                                "type": "string",
                                                "format": "password",
                                                "example": "Passer@123"
                                            },
                                            "role": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "3"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/clients/telephone": {
            "post": {
                "tags": [
                    "SearchClientByTelephone"
                ],
                "summary": "Search Client By Telephone",
                "description": "Récupérer le client par son telephone",
                "operationId": "SearchClientByTelephone",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "telephone"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "785222794"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone"
                                ],
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "761434522"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/clients/{id}": {
            "get": {
                "tags": [
                    "GetClientById"
                ],
                "summary": "Get Client By ID",
                "description": "Récupération du client par son ID",
                "operationId": "GetClientById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Récupérer le client par son ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 17
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération reussie",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}